requirements.txt
pyzmq==25.1.1
numpy==1.23.0
setuptools==69.5.1
torch==2.1.0+cu121
transformers==4.40.0
tqdm==4.66.4
msgpack==1.0.7
